Allow agent to be daemonized
In order to support older SysV style initialization scripts, the Timber
agent must perform daemonization itself if no helper scripts or
executables (such as Brian Clapper's `daemonize`) are available. Since
the agent is written in Go, it does not have native access to the
standard daemonization techniques offered in the C standard library. We
accomplish daemonization instead using the VividCortex godaemon package
which handles creating a new process that is no longer linked to the
executing context.

Running the Timber agent with the daemonization flag should _only_ be
used if there is no other way to accomplish daemonization.

The PID of the executable can be written to a file using the `--pidfile`
commad line option. This is integral to the use of the agent as a
daemon, since it provides a way to shut the daemonized daemon down at a
later point by calling `kill -15 $(cat pidfile)`. (That command sends
the SIGTERM signal to the daemon, which the agent already handles as an
instruction to cleanly shut down).

This also adds the ability for the agent to log to a file provided
as an argument with the `--agent-log-file`.

A caveat of the daemonization process is that it changes the working
directory of the agent to the root of the file system. This means that
any relevant file references will be treated as relevant to the root
rather than relevant to the directory the executable was launched in.

Closes #1
Source of #5