Corrects test for state persistence
The state file's hash of a log file is based on the first 256 bytes of a
file. The file under test had less than 256 bytes when the hash was
generated, and subsequent generations of the hash after data was added
resulted in the file being treated as if it was rotated and the offset
from the state file was not honored.

Since a log file having less than 256 bytes of data is unlikely to
occur, we are not addressing that issue at this time. Instead, the test
has been modified so that the file under test has at least 256 bytes
from the beginning of the test.

Fixes #7