Changes dep manager to golang/dep
In an effort to make managing dependencies easier as well as ensuring
reproducible builds, we are switching from Godeps to golang/dep - the
non-official but pseudo-official Go language dependency manager. All
dependencies are now managed through `Gopkg.toml` which defines
constraints about how to install dependencies. Locking information is
stored in `Gopkg.lock`. Both of these should be committed into the
source repository.

The `vendor` directory is still used to maintain vendorized copies of
dependencies that are suitable for building with.

As part of this change, we forked `github.com/hpcloud/tail` to
`github.com/timberio/tail` in order to integrate an important change
regarding file offsets that had only been made in our vendorized version
of the dependency. (The vendorized dependency was actually
`github.com/influxdata/tail` but `github.com/hpcloud/tail` is the
upstream of that). See github.com/timberio/tail/pull/1 for more details.

Closes #11