Use polling in tests
The CI server runs in a containerized environment that does not fully
support notifying based on file changes. We explicitly use the `Polling`
mechanism instead of depending on these notifications so that tests can
pass.

As part of this change, we also need to increase the timeout for the
expect calls due to the latency with polling.