Send hostname in metadata
This change adds a new header to the forwarding request
(`Timber-Metadata-Override`) which will be used to override data sent in
the log lines encapsulated by the request. The header contains a JSON
object represnting a Timber Log Event based on the JSON schema. The
actual object will not validate against the schema since it is designed
to be merged in to an existing event.

By default, the agent will now retrieve the hostname from the operating
system and send it in this header. The user can also specify the agent
use a specific hostname by setting the `hostname` value in the
configuration file.

Closes #26