Add filename to metadata
Adds the name of the file being tailed to the metadata at
`context.source.file_name`. It also upgrades to v3.0.8 of the LogEvent
schema which supports this field.

This required re-arranging the way that the metadata is built. Since the
metadata needs to be modified on a per-file basis, a base version is now
built and then copies of that are modified before being turned into
JSON.

Closes #35