Fix auto-build on release tag
This change should make CircleCI 2.0 build releases when a new tag
starting with `v` is pushed.

As part of this change:

  1. We switched to using [gox](https://github.com/mitchellh/gox) for
  building cross-platform. Even though cross-platform building is
  natively supported by the golang build tooling, `gox` makes it much
  simpler

  2. We switched to using a [custom Docker
  image](https://github.com/timberio/timber-agent-builder) which
  provides all the necessary dependencies for building.

  3. The CircleCI 2.0 Worfklow usage was removed since it is
  incompatible with the tag-based-build-system.

  4. Releases should now be uploaded to Github automatically using the
  [github-release](https://github.com/c4milo/github-release) tool.

  5. Releases should now be uploaded to S3 automatically.

A lot of these changes were made in the Makefile. While there is still
room for work on the Makefile, it is significantly improved over the
previous iteration.

This fixes #29