Correct Makefile .DEFAULT_GOAL
The default goal (`dist`) was mistakenly set as a dependency of the
goal `.DEFAULT`. The default goal should be a value for the variable
`.DEFAULT_GOAL`. This change corrects that mistake.

It also rearranges the `.PHONY` declarations such that a `.PHONY`
declaration now appears before every phony goal rather than as a single
goal at the top of the file.