Use Docker scratch as base image
Since Go statically links all required dependencies into the binary,
there is no need for additional libraries or an operating system in
order to run the Timber agent inside of a Docker image. To that end, we
are now just copying the Timber agent's Linux distribution folder into
a Docker scratch image.

Closes #28