Merges Elasticbeanstalk configuration files
The Elasticbeanstalk configuration file only works as a single file. It
seems like it is impossible to split the file into two parts (one
handling the user configuration and the other handling the technical
details of installing the Timber agent). This is unfortunate, but we
have to go with what we are given.

Closes #55