Use timberio/agent-builder:2.0.0
This upgrades the Docker image used to build binaries on the continuous
integration server to version 2.0.0 which is based on Debian Jessie
rather than Alpine Linux 3.6. As such, it is based on GNU libc rather
than musl libc. This should resolve the issue with binaries failing to
be loaded at runtime.

Closes #59