Treat AWS EC2 404 as an error
Currently, if the AWS EC2 metadata service is present (in that a
connection to it doesn't time out) we treat this as a success. This is
actually _not_ the intended behavior. If the metadata service is present
but returns a non-successful code, than we should treat this as an
error.

This prevents users from getting large blocks of text with a 404
message as their "public_hostname" value.

Closes #69