Bundle agent log for Elastic Beanstalk
Elastic Beanstalk allows users the option to download the contents of
their instance logs as an archive. In order to be included in this
archive bundle, Elastic Beanstalk needs to be told which additional
files to add through a configuration file under
/opt/elasticbeanstalk/tasks/bundlelogs.d/.

This change adds the /var/log/timber-agent.log file to the bundle.