Allow user to set a default API key
This introduces the concept of a default API key in the configuration
file which can be specified with the `default_api_key` field. If the
default API key is present and a file entry does not have an API key,
the default API key will be used for that file.

This can help significantly when a user has a large number of files but
is not separating the files across different applications by reducing
the reptition of the API key in the configuration file.

Closes #74