Get API key for Elastic Beanstalk from environment
This changes how we handle the API key on AWS Elastic Beanstalk
instances. Previously, the user was expected to add a line that
contained the API key, and we offered advanced instructions on how to
configure the system to change the API key based on the environment.
Now, the extension will rely on fetching the API key from the
environment by default.

This is accomplished using the `Fn::GetOptionSetting` command from
CloudFormation (which is what proceses Elastic Beanstalk configuration
files). The user is expected to set the environment variable
`TIMBER_API_KEY` which is then fetched when the configuration file is
processed. We then use a small `sed` script to inject it into the
top of the configuration file.