Provide functionality through sub commands
This change switches the command line usage for the agent to be based on
sub-commands. At the moment, that means that the flag `--stdin`, which
when present would cause the agent to capture stdin in and when not
present would use the configured file set, has been removed in favor of
dividing functionality into the sub-commands `capture-stdin` and
`capture-files`.

The benefit of this is that it separates out flag usage cleanly; for
example, it did not make sense that `--daemonize` could be used with
`--stdin`. (In theory it could make sense in some edge cases, but in the
majority of cases, allowing it adds needless complexity.)

This change has been supported by a significant refactoring of
funcitonality. As a result, the majority of `run` has been broken up
into smaller, more manageable parts.

User Facing Changes
-------------------

  - `timber-agent` will no longer start the agent, it must be passed a
  sub-command (`capture-stdin` or `capture-files`)
  - Capturing stdin no longer requires a configuration file
  - Capturing stdin can no longer be daemonized

Internal Changes
----------------

  - Parsing a configuration file has been separated from the
  configuration normalization process by splitting `readConfig` into
  `parseConfig` and `normalizeConfig`
  - Configuration validation has been split up to handle configuration
  for capturing stdin and configuration for capturing files seaprately.
  `validateConfig` has been replaced by `validateConfigFiles` and
  `validConfigStdin`.
  - Metadata is now passed to `Forward` as a string instead of a byte
  array. This should result in an imperceptible speed increase the byte
  array does not need to be coerced to a string on each HTTP call.
  - `NewReaderTailer` now returns a `*ReaderTailer` instead of
  `ReaderTailer`. This allows the return to act as a `Tailer`
  implementation, since `Lines` expects to have a `*ReaderTailer` as the
  receiver.
  - `run` has been split up into many parts. There are now two separate
  entry points: `runCaptureFiles` and `runCaptureStdin`.
  `runCaptureFiles` is used for capturing log data from files and
  `runCaptureStdin` is used for capturing log data over STDIN.

After looking through the code that handles capturing STDIN and
capturing files, it became apparent that the major difference between
the two was in which configuration parameters they each required and how
they produced the tailing struct to be used in the batching go-routine.
Beyond these differences, the majority of the code was the same. This
made it possible to factor out the shared pieces into their own
function which could be called by both paths. The relevant command would
then only differ in the way it built up the sources to tail.

The common setup for both functions has been moved to
`commonCaptureRunSetup` and the teardown work has been moved to
`commonCaptureRunTeardown`.

`commonCaptureRunSetup` returns a pointer to a `commonCaptureRunContext`
which formalizes the shared parameters of a capture command. It contains
a configured instance of `*log.Logger`, a `*Config` read from the config
file (or empty if not provided in the case of `runCaptureStdin`), the
pidfile path, the logfile path, the quit channel, general metadata, and
an empty array of sources.

Once control is given back to the caller, the caller is responsible for
building the list of sources its own way. A source is composed of an API
key, metadata (as a string), and something implementing the Tailer
interface. In `runCaptureStdin` a single source is built consisting of a
`*ReaderTailer`. In `runCaptureFiles`, a source is build for each file
consisting of a `*FileTailer`.

These sources are then assigned to the `Sources` field on the
`commonCaptureRunContext`. The run context is passed to `commonTailRun`
which loops through the sources and create a buffer channel for each
which is given to a batching go routine before being passed to
`Forward`. A wait group is formed that waits until `Forward` has
returned for all sources.

Once the wait group returns, `commonTailRunTeardown` is called which
performs some common cleanup. (Currently this is limited to removing the
PID file).

Closes #93

Closes #14