Update examples with sub-command usage
This change updates example scripts to use sub-commands instead of the
`--stdin` flag.

As part of this, the sysv-epel example script now uses `--pidfile`
instead of `--pid-file`. The erroneous flag usage was causing errors for
some users.

It also adds an `UPGRADING.md` document which provides an overview of
changes necessary to upgrade.

Closes #97

Closes #101