Move authentication string builder out of loop
I moved the formatting code for the `Authorization` header's value (the
authentication string) to be above the loop its used in. There is no
reason that a new value should be built each time. It's the same string.
This reduces resources usage by a such a ridiculous amount I don't think
it's even measureable.