Reference LogEvent field structs by pointer
There has been an ongoing issue where the AWS EC2 metadata that gets
sent with the header is actually invalid because it contains empty JSON
objects which fail to validate against the schema. The reason for this
is that Go's JSON encoding treats structs as always-present, so the
`omitempty` directive has no bearing. In order for the empty object to
be elided, the field must have a nil value. By switching the LogEvent
struct (and all its associated structs) to use pointers for field
structs instead of the struct value, the `omitempty` takes appropriate
effect.

This introduces more complexity in the codebase. The field must now be
checked for a pointer before descendant fields can be appropriately
assigned or accessed. Helper functions are available, but this will
require testing before releases to ensure no errant pointer issues make
it into releases.

Closes #98
Closes #123