Avoid logger.fatal on agent forwarding errors
In attempt to keep the agent running and avoid hard crashes, we return
client errors from forwarding when encountered. This wil cause the
goroutine for that file to exit, but keep the agent running serving
other behaving files where client config is correct.

On server errors, unless we drop the logs, we exit and assume the
process will restart and retry the failures. A more robust solution
should be introduced be will not be at this time.

Closes #154