Add fixes and docs for Tectonic hostname error
The Timber Agent gathers host information via the /etc/hostname file,
which Tectonic worker nodes do not have. As such, Tectonic specific
manifests have been created which do not mount /etc/hostname. While this
means that the reported hostname will be the name of the Timber Agent
container, it is better than not shipping logs.